# webharvest-openapi-spec.yaml
openapi: 3.0.3
info:
  title: WebHarvest API
  description: |
    Self-hosted web scraping platform with AI integration. WebHarvest provides powerful web crawling, scraping, and content extraction capabilities with built-in caching, batch processing, and AI-powered content analysis.

    ## Key Features
    - **Web Scraping**: Extract content from single URLs in various formats (Markdown, HTML, screenshots)
    - **Web Crawling**: Discover and crawl entire websites with intelligent depth control
    - **Batch Processing**: Process multiple URLs simultaneously for efficient data collection
    - **Site Mapping**: Fast URL discovery without content extraction
    - **Content Caching**: Intelligent caching system to reduce redundant requests
    - **MCP Integration**: Model Context Protocol support for AI workflows
    - **Project Management**: Organize and manage crawling projects
    - **Real-time Monitoring**: Track job progress and health status
    - **OpenWebUI Integration**: Sync content to knowledge collections

    ## Authentication
    All API endpoints (except health checks) require authentication using API keys.
    Include the API key in the Authorization header as a Bearer token:
    
    ```
    Authorization: Bearer wh_your_api_key_here
    ```

    ## Rate Limiting
    - Default rate limit: 60 requests per minute per API key
    - Configurable per API key
    - 503 status returned when rate limit exceeded

    ## Response Formats
...

  version: 1.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /v2/crawl:
    post:
      summary: Start Crawl
      description: |
        Start a crawl job to discover and scrape an entire website.
        Returns immediately with a job ID for monitoring progress.
        
        **Features:**
        - Configurable crawl depth and page limits
        - Include/exclude path patterns using regex
        - Sitemap discovery and processing
        - Domain and subdomain control
        - Rate limiting and concurrency control
        - Webhook notifications for completion
      tags: [Crawling]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrawlRequest'
            examples:
              basic_crawl:
                summary: Basic website crawl
                value:
                  url: "https://docs.example.com"
                  maxDiscoveryDepth: 3
                  limit: 100
              advanced_crawl:
                summary: Advanced crawl with filtering
                value:
                  url: "https://docs.example.com"
                  maxDiscoveryDepth: 5
                  limit: 500
                  includePaths: ["/docs/.*", "/api/.*"]
                  excludePaths: ["/docs/old/.*"]
                  crawlEntireDomain: true
                  maxConcurrency: 3
      responses:
        '200':
          description: Crawl job started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrawlStartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /v2/crawl/{crawl_id}:
    get:
      summary: Get Crawl Status
      description: |
        Get the status and results of a crawl job.
        Returns progress information and scraped pages for completed jobs.
      tags: [Crawling]
      parameters:
        - name: crawl_id
          in: path
          required: true
          description: Unique identifier of the crawl job
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Crawl status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrawlStatusResponse'
        '404':
          description: Crawl job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    
    delete:
      summary: Cancel Crawl
      description: Cancel a running or queued crawl job
      tags: [Crawling]
      parameters:
        - name: crawl_id
          in: path
          required: true
          description: Unique identifier of the crawl job to cancel
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Crawl job canceled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrawlCancelResponse'
        '404':
          description: Crawl job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v2/batch/scrape:
    post:
      summary: Start Batch Scrape
      description: |
        Start a batch scraping job to process multiple URLs simultaneously.
        Efficient for scraping multiple pages with consistent configuration.
        
        **Features:**
        - Concurrent processing with configurable concurrency
        - Invalid URL filtering and reporting
        - Webhook notifications for completion
        - Progress tracking and error reporting
      tags: [Batch]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchScrapeRequest'
            examples:
              basic_batch:
                summary: Basic batch scraping
                value:
                  urls:
                    - "https://example.com/page1"
                    - "https://example.com/page2"
                    - "https://example.com/page3"
                  maxConcurrency: 5
              advanced_batch:
                summary: Batch with custom options
                value:
                  urls:
                    - "https://example.com/page1"
                    - "https://example.com/page2"
                  ignoreInvalidURLs: true
                  maxConcurrency: 10
                  scrapeOptions:
                    formats: ["markdown", "links"]
                    onlyMainContent: true
      responses:
        '200':
          description: Batch scrape job started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchScrapeStartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /v2/batch/scrape/{batch_id}:
    get:
      summary: Get Batch Status
      description: Get the status and results of a batch scraping job
      tags: [Batch]
      parameters:
        - name: batch_id
          in: path
          required: true
          description: Unique identifier of the batch job
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Batch status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchStatusResponse'
        '404':
          description: Batch job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key authentication using Bearer tokens.
        Format: `Bearer wh_your_api_key_here`
        
        API keys must start with 'wh_' prefix and be obtained from your WebHarvest dashboard.
  responses:
    BadRequest:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Bad Request"
            message: "Invalid URL format"
    Unauthorized:
      description: Unauthorized - Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Unauthorized"
            message: "Invalid or missing API key"
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Rate Limited"
            message: "API rate limit exceeded. Try again later."
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Internal Server Error"
            message: "An unexpected error occurred"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Detailed error message
      required: [success, error, message]

    # Scraping Schemas
...
    PageMetadata:
      type: object
      properties:
        title:
          type: string
          description: Page title
        description:
          type: string
          description: Meta description
        language:
          type: string
          description: Detected language
        sourceURL:
          type: string
          format: uri
          description: Canonical source URL
        statusCode:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: Error message if scraping failed

    # Crawling Schemas
    CrawlRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
          description: Starting URL for crawl
          example: "https://docs.example.com"
        excludePaths:
          type: array
          items:
            type: string
          description: Regex patterns for URLs to exclude
          example: ["/admin/.*", "/private/.*"]
        includePaths:
          type: array
          items:
            type: string
          description: Regex patterns for URLs to include
          example: ["/docs/.*", "/api/.*"]
        maxDiscoveryDepth:
          type: integer
          default: 10
          minimum: 1
          maximum: 20
          description: Maximum crawl depth
        sitemap:
          type: string
          enum: [include, ignore, only]
          default: include
          description: How to handle sitemap discovery
        ignoreQueryParameters:
          type: boolean
          default: false
          description: Ignore query parameters in URL normalization
        limit:
          type: integer
          default: 5000
          minimum: 1
          maximum: 100000
          description: Maximum number of pages to crawl
        crawlEntireDomain:
          type: boolean
          default: false
          description: Allow crawling entire domain (not just subdirectories)
        allowExternalLinks:
          type: boolean
          default: false
          description: Follow links to external domains
        allowSubdomains:
          type: boolean
          default: false
          description: Allow crawling subdomains
        delay:
          type: integer
          default: 250
          minimum: 0
          maximum: 10000
          description: Delay between requests in milliseconds
        maxConcurrency:
          type: integer
          default: 5
          minimum: 1
          maximum: 20
          description: Maximum concurrent requests
        webhook:
          type: object
          properties:
            url:
              type: string
              format: uri
              description: Webhook URL for completion notification
            headers:
              type: object
              additionalProperties:
                type: string
              description: Custom headers for webhook
          description: Webhook configuration for job completion
        scrapeOptions:
          type: object
          description: Options to apply to each scraped page
          properties:
            formats:
              type: array
              items:
                type: string
                enum: [markdown, html, rawHtml, links, images]
            onlyMainContent:
              type: boolean
              default: true
      required: [url]
    CrawlStartResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        id:
          type: string
          format: uuid
          description: Unique crawl job identifier
        url:
          type: string
          description: URL to check crawl status
          example: "/v2/crawl/550e8400-e29b-41d4-a716-446655440000"
      required: [success, id, url]
    CrawlStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        status:
          type: string
          enum: [queued, running, completed, failed, canceled]
          description: Current job status
        total:
          type: integer
          description: Total URLs discovered
        completed:
          type: integer
          description: Number of URLs successfully processed
        failed:
          type: integer
          description: Number of URLs that failed processing
        data:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
                format: uri
              markdown:
                type: string
              metadata:
                $ref: '#/components/schemas/PageMetadata'
          description: Array of scraped page results
      required: [success, status, total, completed, failed, data]
    CrawlCancelResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Crawl job canceled"
      required: [success, message]

    # Mapping Schemas
...
    BatchScrapeRequest:
      type: object
      properties:
        urls:
          type: array
          items:
            type: string
            format: uri
          minItems: 1
          maxItems: 10000
          description: Array of URLs to scrape
          example: ["https://example.com/page1", "https://example.com/page2"]
        ignoreInvalidURLs:
          type: boolean
          default: false
          description: Continue processing if some URLs are invalid
        maxConcurrency:
          type: integer
          default: 10
          minimum: 1
          maximum: 50
          description: Maximum concurrent scraping operations
        scrapeOptions:
          type: object
          description: Scraping options to apply to all URLs
          properties:
            formats:
              type: array
              items:
                type: string
                enum: [markdown, html, rawHtml, links, images]
            onlyMainContent:
              type: boolean
              default: true
        webhook:
          type: object
          properties:
            url:
              type: string
              format: uri
            headers:
              type: object
              additionalProperties:
                type: string
          description: Webhook configuration for completion notification
      required: [urls]
    BatchScrapeStartResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        id:
          type: string
          format: uuid
          description: Unique batch job identifier
        url:
          type: string
          description: URL to check batch status
        invalidURLs:
          type: array
          items:
            type: string
          description: List of invalid URLs that were excluded
      required: [success, id, url]
    BatchStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        status:
          type: string
          enum: [queued, running, completed, failed]
        total:
          type: integer
          description: Total number of URLs in batch
        completed:
          type: integer
          description: Number of URLs successfully processed
        failed:
          type: integer
          description: Number of URLs that failed
        data:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
                format: uri
              markdown:
                type: string
              metadata:
                $ref: '#/components/schemas/PageMetadata'
          description: Array of batch processing results
      required: [success, status, total, completed, failed, data]

    # MCP (Model Context Protocol) Schemas
...